Updated computational speed and directories
Modified grid algorithm and population calibration to improve computational speed and new way to choose file directories

+++ runner.py
@@ -1,47 +1,66 @@
 # Pulls all the other functions together to make magic!
-#
-# Author: Christopher Arderne
-# Date: 26 November 2016
-# Python version: 3.5
+# Author: KTH dESA Last modified by Alexandros Korkovelos
+# Date: 23 November 2018
+# Python version: 3.7
+
+# Modified grid algorithm and population calibration to improve computational speed
 
 import os
-from pyonsset.onsset import *
+from onsset import *
+import tkinter as tk
+from tkinter import filedialog, messagebox
 
-os.chdir('..')
-os.chdir('db')
+root = tk.Tk()
+root.withdraw()
+root.attributes("-topmost", True)
 
-specs_path = str(input('Enter the name of the specs file: '))
+messagebox.showinfo('OnSSET', 'Open the specs file')
+specs_path = filedialog.askopenfilename()
+
 specs = pd.read_excel(specs_path, index_col=0)
+
+coordinate_units = 1000 # 1000 if coordinates are in m, 1 if coordinates are in km
+
+# specs_directory = str(input('Enter the directory of the specs file: '))
+# os.chdir(specs_directory)
+# specs_path = str(input('Enter the name of the specs file: '))
+#
+# try:
+#     specs = pd.read_excel(specs_path, index_col=0)
+# except FileNotFoundError:
+#     specs = pd.read_excel(str(specs_path + '.xlsx'), index_col=0)
 
 countries = str(input('countries: ')).split()
 countries = specs.index.tolist() if 'all' in countries else countries
 
-choice = int(input('Enter 1 to split, 2 to prep, 3 to run a scenario: '))
+choice = int(input('1 to calibrate and prep, 2 to run a scenario: '))
 
-if choice == 1:
-    settlements_csv = str(input('Enter the name of the file containing all countries: '))
-    base_dir = str(input('Enter the base file directory to save the split countries: '))
+if choice == 0:
+    messagebox.showinfo('OnSSET', 'Open the csv file with GIS data')
+    settlements_csv = filedialog.askopenfilename()
+    messagebox.showinfo('OnSSET', 'Select the folder to save split countries')
+    base_dir = filedialog.asksaveasfilename()
 
     print('\n --- Splitting --- \n')
-
-    try:
-        os.makedirs(base_dir)
-    except FileExistsError:
-        pass
 
     df = pd.read_csv(settlements_csv)
 
     for country in countries:
         print(country)
-        df.loc[df[SET_COUNTRY] == country].to_csv(os.path.join(base_dir, '{}.csv'.format(country)), index=False)
+        df.loc[df[SET_COUNTRY] == country].to_csv(base_dir + '.csv', index=False)
 
-elif choice == 2:
-    base_dir = str(input('Enter the base file directory containing separated countries (files will be overwritten): '))
+elif choice == 1:
+    messagebox.showinfo('OnSSET', 'Open the csv file containing the extracted GIS data')
+    base_dir = filedialog.askopenfilename()
+    messagebox.showinfo('OnSSET', 'Browse to result folder and name the calibrated file')
+    output_dir = filedialog.asksaveasfilename()
+
     print('\n --- Prepping --- \n')
 
     for country in countries:
         print(country)
-        settlements_in_csv = os.path.join(base_dir, '{}.csv'.format(country))
+        settlements_in_csv = base_dir  # os.path.join(base_dir, '{}.csv'.format(country))
+        settlements_out_csv = output_dir + '.csv'  # os.path.join(output_dir, '{}.csv'.format(country))
 
         onsseter = SettlementProcessor(settlements_in_csv)
 
@@ -75,15 +94,18 @@ elif choice == 2:
         specs.loc[country, SPE_ELEC_MODELLED] = elec_modelled
         specs.loc[country, SPE_POP_CUTOFF1] = pop_cutoff
         specs.loc[country, SPE_POP_CUTOFF2] = pop_cutoff2
+        specs.loc[country, SPE_URBAN_MODELLED] = urban_modelled
+        specs.loc[country, SPE_URBAN_CUTOFF] = urban_cutoff
 
-        specs.to_excel(specs_path)
-        onsseter.df.to_csv(settlements_in_csv, index=False)
+        try:
+            specs.to_excel(specs_path)
+        except ValueError:
+            specs.to_excel(specs_path + '.xlsx')
 
-elif choice == 3:
-    base_dir = str(input('Enter the base file directory containing separated and prepped countries: '))
-    output_dir = str(input('Enter the output directory (can contain multiple runs): '))
+        onsseter.df.to_csv(settlements_out_csv, index=False)
 
-    wb_tiers_all = {1: 7.738, 2: 43.8, 3: 160.6, 4: 423.4, 5: 598.6}
+elif choice == 2:
+    wb_tiers_all = {1: 8, 2: 44, 3: 160, 4: 423, 5: 598}
     print("""\nWorld Bank Tiers of Electricity Access
           1: {} kWh/person/year
           2: {} kWh/person/year
@@ -96,21 +118,24 @@ elif choice == 3:
 
     diesel_high = True if 'y' in input('Use high diesel value? <y/n> ') else False
     diesel_tag = 'high' if diesel_high else 'low'
-    do_combine = True if 'y' in input('Combine countries into a single file? <y/n> ') else False
+
+    messagebox.showinfo('OnSSET', 'Open the csv file with calibrated GIS data')
+    base_dir = filedialog.askopenfilename()
+    messagebox.showinfo('OnSSET', 'Browse to result folder and name the scenario to save outputs')
+    output_dir = filedialog.asksaveasfilename()
+
+    # Uncomment row below if running multiple countries/regions
+    do_combine = False
+    # do_combine = True if 'y' in input('Combine countries into a single file? <y/n> ') else False
 
     print('\n --- Running scenario --- \n')
-
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        pass
 
     for country in countries:
         # create country_specs here
         print(' --- {} --- {} --- {} --- '.format(country, wb_tier_urban, diesel_tag))
-        settlements_in_csv = os.path.join(base_dir, '{}.csv'.format(country))
-        settlements_out_csv = os.path.join(output_dir, '{}_{}_{}.csv'.format(country, wb_tier_urban, diesel_tag))
-        summary_csv = os.path.join(output_dir, '{}_{}_{}_summary.csv'.format(country, wb_tier_urban, diesel_tag))
+        settlements_in_csv = base_dir
+        settlements_out_csv = output_dir + '.csv'
+        summary_csv = output_dir + 'summary.csv'
 
         onsseter = SettlementProcessor(settlements_in_csv)
 
@@ -212,16 +237,28 @@ elif choice == 3:
         grid_lcoes_urban = grid_calc.get_grid_table(energy_per_hh_urban, num_people_per_hh_urban,
                                                     max_grid_extension_dist)
         onsseter.run_elec(grid_lcoes_rural, grid_lcoes_urban, grid_price,
-                          existing_grid_cost_ratio, max_grid_extension_dist)
+                          existing_grid_cost_ratio, max_grid_extension_dist, coordinate_units)
 
         onsseter.results_columns(mg_hydro_calc, mg_wind_calc, mg_pv_calc, sa_pv_calc,
                                  mg_diesel_calc, sa_diesel_calc, grid_calc)
 
         summary = onsseter.calc_summaries()
         summary.name = country
-        summary.to_csv(summary_csv, header=True)
 
-        onsseter.df.to_csv(settlements_out_csv, index=False)
+        try:
+            summary.to_csv(summary_csv, header=True)
+        except PermissionError:
+            if 'y' in input('Summary file open. Close it and enter "y" to overwrite (or rename the open file first)'):
+                summary.to_csv(summary_csv, header=True)
+            else:
+                pass
+        try:
+            onsseter.df.to_csv(settlements_out_csv, index=False)
+        except PermissionError:
+            if 'y' in input('Output csv file open. Close it and enter "y" to overwrite (or rename open file first)'):
+                onsseter.df.to_csv(settlements_out_csv, index=False)
+            else:
+                pass
 
     if do_combine:
         print('\n --- Combining --- \n')
@@ -239,5 +276,9 @@ elif choice == 3:
                                              squeeze=True, index_col=0)
 
         print('saving csv')
-        df_base.to_csv(os.path.join(output_dir, '{}_{}.csv'.format(wb_tier_urban, diesel_tag)), index=False)
-        summaries.to_csv(os.path.join(output_dir, '{}_{}_summary.csv'.format(wb_tier_urban, diesel_tag)))
+        df_base.to_csv(os.path.join(output_dir, '{}_{}_{}.csv'.format(wb_tier_urban,
+                                                                      wb_tier_rural, diesel_tag)), index=False)
+        summaries.to_csv(os.path.join(output_dir, '{}_{}_{}_summary.csv'.format(wb_tier_urban,
+                                                                                wb_tier_rural, diesel_tag)))
+
+    logging.info('Scenario run finished')
